refactor the details page into customizations
