+++ components/Wizard/steps/imageOutput.js
@@ -1,6 +1,7 @@
+/* eslint-disable no-unused-vars */
 import React from "react";
 import { Button, Popover } from "@patternfly/react-core";
-import { OutlinedQuestionCircleIcon } from "@patternfly/react-icons";
+import { HelpIcon, OutlinedQuestionCircleIcon } from "@patternfly/react-icons";
 
 import { componentTypes, validatorTypes } from "@data-driven-forms/react-form-renderer";
 import nextStepMapper from "./imageOutputStepMapper";
@@ -135,6 +136,86 @@ export default {
       condition: {
         when: "image-output-type",
         is: "oci",
+      },
+    },
+    {
+      component: "text-field-custom",
+      name: "image-size",
+      className: "pf-u-w-50",
+      type: "number",
+      label: "Image size",
+      labelIcon: (
+        <Popover
+          bodyContent={
+            <>
+              Set the size that you want the image to be when instantiated. The total package size and target
+              destination of your image should be considered when setting the image size.
+            </>
+          }
+          aria-label="Image size help"
+        >
+          <Button variant="plain" aria-label="Image size help">
+            <HelpIcon />
+          </Button>
+        </Popover>
+      ),
+      helperText: "Minimum image size is 2GB",
+      initializeOnMount: true,
+      initialValue: 2,
+      isRequired: true,
+      autoFocus: true,
+      condition: {
+        when: "image-output-type",
+        is: ["", "fedora-iot-commit", "edge-commit", "edge-container", "edge-installer", "image-installer"],
+        notMatch: true,
+      },
+      validate: [
+        {
+          type: validatorTypes.REQUIRED,
+        },
+        {
+          type: validatorTypes.MIN_NUMBER_VALUE,
+          includeThreshold: true,
+          value: 2,
+          message: "Minimum image size is 2GB",
+        },
+      ],
+      resolveProps: (props, { meta, input }, formOptions) => {
+        const imageType = formOptions.getState().values["image-output-type"];
+        if (imageType === "ami") {
+          return {
+            initialValue: 6,
+            helperText: "Minimum image size is 6GB",
+            validate: [
+              {
+                type: validatorTypes.REQUIRED,
+              },
+              {
+                type: validatorTypes.MIN_NUMBER_VALUE,
+                includeThreshold: true,
+                value: 6,
+                message: "Minimum image size is 6GB",
+              },
+            ],
+          };
+        }
+        if (imageType === "edge-simplified-installer") {
+          return {
+            initialValue: 10,
+            helperText: "Minimum image size is 10GB",
+            validate: [
+              {
+                type: validatorTypes.REQUIRED,
+              },
+              {
+                type: validatorTypes.MIN_NUMBER_VALUE,
+                includeThreshold: true,
+                value: 6,
+                message: "Minimum image size is 10GB",
+              },
+            ],
+          };
+        }
       },
     },
   ],
