+++ components/Wizard/steps/index.js
@@ -8,6 +8,6 @@ export { default as ociDest } from "./ociDest";
 export { default as vmwareAuth } from "./vmwareAuth";
 export { default as vmwareDest } from "./vmwareDest";
 export { default as ostreeSettings } from "./ostreeSettings";
-export { default as details } from "./details";
+export { default as customizations } from "./customizations";
 export { default as packages } from "./packages";
 export { default as review } from "./review";
