+++ components/Wizard/steps/ociDest.js
@@ -7,7 +7,7 @@ export default {
   title: "Destination",
   name: "oci-dest",
   substepOf: "Upload to OCI",
-  nextStep: "details",
+  nextStep: "customizations",
   fields: [
     {
       component: "text-field-custom",
