+++ components/Wizard/steps/ostreeSettings.js
@@ -6,7 +6,7 @@ import { HelpIcon } from "@patternfly/react-icons";
 export default {
   title: "OSTree settings",
   name: "ostree-settings",
-  nextStep: "details",
+  nextStep: "customizations",
   fields: [
     {
       component: "text-field-custom",
